test/webtests/http/tests/devtools/components/common-prefix-expected.txt

Tests TextUtils.TextUtils.commonPrefix.


Running: testSimple
prefix/

Running: testFilepaths
c:\win32\

Running: testNoSharedPrefix


Running: testNoStrings
